+++ src/rpc/expel.rs
@@ -1,13 +1,13 @@
 use std::time::Duration;
 
-use crate::instance::{Instance, StateVariant::*};
+use crate::has_states;
+use crate::instance::StateVariant::*;
 use crate::rpc;
 use crate::rpc::update_instance::handle_update_instance_request_and_wait;
-use crate::traft::error::IdOfInstance;
 use crate::traft::Result;
 use crate::traft::{error::Error, node};
-
-const TIMEOUT: Duration = Duration::from_secs(10);
+use tarantool::error::BoxError;
+use tarantool::error::TarantoolErrorCode;
 
 crate::define_rpc_request! {
     /// Submits a request to expel the specified instance. If successful
@@ -34,14 +34,21 @@ crate::define_rpc_request! {
             });
         }
 
-        let instance = node.storage.instances.by_uuid(&req.instance_uuid)?;
-        let Some(Instance { name, .. }) = instance else {
-            return Err(Error::NoSuchInstance(IdOfInstance::Uuid(req.instance_uuid)));
-        };
+        let topology_ref = node.topology_cache.get();
+        let instance = topology_ref.instance_by_uuid(&req.instance_uuid)?;
+        if has_states!(instance, * -> Expelled) {
+            // Idempotency
+            return Ok(Response {});
+        }
 
-        let req = rpc::update_instance::Request::new(name, req.cluster_name)
+        let timeout = req.timeout;
+        let req = rpc::update_instance::Request::new(instance.name.clone(), req.cluster_name)
             .with_target_state(Expelled);
-        handle_update_instance_request_and_wait(req, TIMEOUT)?;
+
+        // Must not hold this reference across yields
+        drop(topology_ref);
+
+        handle_update_instance_request_and_wait(req, timeout)?;
 
         Ok(Response {})
     }
@@ -52,12 +59,14 @@ crate::define_rpc_request! {
     pub struct Request {
         pub cluster_name: String,
         pub instance_uuid: String,
+        pub timeout: Duration,
     }
 
     pub struct Response {}
 }
 
 pub mod redirect {
+    use super::*;
     use ::tarantool::fiber;
 
     use crate::rpc::network_call_to_leader;
@@ -66,7 +75,31 @@ pub mod redirect {
     crate::define_rpc_request! {
         fn proc_expel_redirect(req: Request) -> Result<Response> {
             let Request(req_to_leader) = req;
+
+            let deadline = fiber::clock().saturating_add(req_to_leader.timeout);
             fiber::block_on(network_call_to_leader(crate::proc_name!(super::proc_expel), &req_to_leader))?;
+
+            let node = node::global()?;
+            let instance_uuid = &req_to_leader.instance_uuid;
+            loop {
+                let topology_ref = node.topology_cache.get();
+                let instance = topology_ref.instance_by_uuid(instance_uuid)?;
+                if has_states!(instance, Expelled -> *) {
+                    break
+                }
+
+                let now = fiber::clock();
+                if now > deadline {
+                    return Err(BoxError::new(TarantoolErrorCode::Timeout, "expel confirmation didn't arrive in time").into());
+                }
+
+                // Must not hold this reference across yields
+                drop(topology_ref);
+
+                let timeout = deadline.duration_since(now);
+                _ = node.wait_index(node.get_index() + 1, timeout);
+            }
+
             Ok(Response {})
         }